+++ src/components/query/index.tsx
@@ -1,12 +1,9 @@
 import { FunctionComponent } from 'react';
 
-import { useRouter } from 'next/router';
-
 import { makeStyles, Theme } from '@material-ui/core';
 
-import useConnection from '../../hooks/useConnection';
+import useConnectionContext from '../../hooks/useConnectionContext';
 import useQuery from '../../hooks/useQuery';
-import useDashboard from '../../hooks/useDashboard';
 
 import Form from './_form';
 import Results from '../results';
@@ -32,28 +29,14 @@ type QueryProps = {};
 
 const Query: FunctionComponent<QueryProps> = () => {
   const classes = useStyles();
-  const router = useRouter();
 
-  const { connection } = useConnection();
-
-  const onPin = () => {
-    router.push('/dashboard');
-  };
-
-  const { onPush } = useDashboard(connection ? connection.name : '');
+  const { connection } = useConnectionContext();
 
   const { query, setQuery, error, isRunning, results } = useQuery(connection);
 
   return (
     <div className={classes.root}>
-      <Form
-        setQuery={setQuery}
-        query={query}
-        error={error}
-        isRunning={isRunning}
-        onDashboardPush={onPush}
-        onPin={onPin}
-      />
+      <Form setQuery={setQuery} query={query} error={error} isRunning={isRunning} />
       <div className={classes.results}>
         <Results query={query} results={results} />
       </div>
