+++ src/hooks/useConnection/index.ts
@@ -1,11 +1,18 @@
-import { useContext } from 'react';
+import { useContext, useEffect, useState } from 'react';
+
+import { Connection } from '../../models';
 
 import context from '../../contexts/connection';
 
-const useConnection = () => {
-  const { connection, connections, onChange, onSave } = useContext(context);
+const useConnection = (name: string) => {
+  const { connections } = useContext(context);
+  const [connection, setConnection] = useState<Connection>();
 
-  return { connection, connections, onChange, onSave };
+  useEffect(() => {
+    setConnection(connections.filter(c => c.name === name)[0]);
+  }, [name]);
+
+  return connection;
 };
 
 export default useConnection;
