+++ src/pages/dashboard/index.tsx
@@ -1,0 +1,9 @@
+import { FunctionComponent } from 'react';
+
+import Dashboard from '../../components/dashboard';
+
+const DashboardPage: FunctionComponent = () => {
+  return <Dashboard />;
+};
+
+export default DashboardPage;
